chore(checkpoint): preserve progress at turn 150
Sybra checkpointed this run at the per-run turn ceiling (150 turns) before handing off to a fresh agent.
