+++ pkg/plugins/policies/meshfaultinjection/api/v1alpha1/deprecated.go
@@ -2,13 +2,8 @@ package v1alpha1
 
 import (
 	"github.com/kumahq/kuma/v3/pkg/plugins/policies/core/jsonpatch/validators"
-	"github.com/kumahq/kuma/v3/pkg/util/pointer"
 )
 
 func (t *MeshFaultInjectionResource) Deprecations() []string {
-	var deprecations []string
-	if len(pointer.Deref(t.Spec.From)) > 0 {
-		deprecations = append(deprecations, "'from' field is deprecated, use 'rules' with SPIFFE-based 'matches' instead")
-	}
-	return append(deprecations, validators.TopLevelTargetRefDeprecations(t.Spec.TargetRef)...)
+	return validators.TopLevelTargetRefDeprecations(t.Spec.TargetRef)
 }
