+++ pkg/plugins/policies/meshfaultinjection/plugin/v1alpha1/plugin.go
@@ -3,27 +3,19 @@ package v1alpha1
 import (
 	envoy_listener "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
 
-	mesh_proto "github.com/kumahq/kuma/v3/api/mesh/v1alpha1"
-	"github.com/kumahq/kuma/v3/pkg/core"
 	core_meta "github.com/kumahq/kuma/v3/pkg/core/metadata"
 	core_plugins "github.com/kumahq/kuma/v3/pkg/core/plugins"
 	core_mesh "github.com/kumahq/kuma/v3/pkg/core/resources/apis/mesh"
 	core_xds "github.com/kumahq/kuma/v3/pkg/core/xds"
-	util "github.com/kumahq/kuma/v3/pkg/plugins/policies/core/egress"
 	"github.com/kumahq/kuma/v3/pkg/plugins/policies/core/matchers"
 	core_rules "github.com/kumahq/kuma/v3/pkg/plugins/policies/core/rules"
 	policies_xds "github.com/kumahq/kuma/v3/pkg/plugins/policies/core/xds"
 	api "github.com/kumahq/kuma/v3/pkg/plugins/policies/meshfaultinjection/api/v1alpha1"
 	plugin_xds "github.com/kumahq/kuma/v3/pkg/plugins/policies/meshfaultinjection/plugin/xds"
-	"github.com/kumahq/kuma/v3/pkg/util/pointer"
 	xds_context "github.com/kumahq/kuma/v3/pkg/xds/context"
-	"github.com/kumahq/kuma/v3/pkg/xds/envoy/names"
 )
 
-var (
-	_   core_plugins.EgressPolicyPlugin = &plugin{}
-	log                                 = core.Log.WithName("MeshFaultInjection")
-)
+var _ core_plugins.EgressPolicyPlugin = &plugin{}
 
 type plugin struct{}
 
@@ -43,7 +35,10 @@ func (p plugin) EgressMatchedPolicies(tags map[string]string, resources xds_cont
 
 func (p plugin) Apply(rs *core_xds.ResourceSet, ctx xds_context.Context, proxy *core_xds.Proxy) error {
 	if proxy.ZoneEgressProxy != nil {
-		return applyToEgress(rs, proxy)
+		// MeshFaultInjection no longer supports targeting the legacy ExternalService resource via
+		// 'from' on zone egress. MeshExternalService fault injection on zone egress is applied
+		// through the Dataplane path below using rules-based SNI matches (applyToZoneProxyListener).
+		return nil
 	}
 
 	if proxy.Dataplane == nil {
@@ -85,25 +80,15 @@ func applyToInbounds(
 
 		inboundRules, ok := fromRules.InboundRules[listenerKey]
 		if !ok || len(inboundRules) == 0 {
-			//nolint:staticcheck // SA1019 Backward compatibility: fallback to old Rules format if InboundRules not present
-			rules, ok := fromRules.Rules[listenerKey]
-			if !ok {
-				continue
-			}
+			continue
+		}
 
+		switch protocol {
+		case core_meta.ProtocolHTTP, core_meta.ProtocolHTTP2, core_meta.ProtocolGRPC:
+			configurer := plugin_xds.Configurer{Rules: inboundRules}
 			for _, filterChain := range listener.FilterChains {
-				if err := configure(rules, filterChain, protocol); err != nil {
+				if err := configurer.ConfigureHttpListener(filterChain); err != nil {
 					return err
-				}
-			}
-		} else {
-			switch protocol {
-			case core_meta.ProtocolHTTP, core_meta.ProtocolHTTP2, core_meta.ProtocolGRPC:
-				configurer := plugin_xds.Configurer{Rules: inboundRules}
-				for _, filterChain := range listener.FilterChains {
-					if err := configurer.ConfigureHttpListener(filterChain); err != nil {
-						return err
-					}
 				}
 			}
 		}
@@ -169,66 +154,3 @@ func applyToZoneProxyListener(
 	return nil
 }
 
-func applyToEgress(rs *core_xds.ResourceSet, proxy *core_xds.Proxy) error {
-	listeners := policies_xds.GatherListeners(rs)
-	if listeners.Egress == nil {
-		log.V(1).Info("skip applying MeshFaultInjection, Egress has no listener",
-			"proxyName", proxy.ZoneEgressProxy.ZoneEgressResource.GetMeta().GetName(),
-		)
-		return nil
-	}
-	for _, resource := range proxy.ZoneEgressProxy.MeshResourcesList {
-		for _, es := range resource.ExternalServices {
-			meshName := resource.Mesh.GetMeta().GetName()
-			esName, ok := es.Spec.GetTags()[mesh_proto.ServiceTag]
-			if !ok {
-				continue
-			}
-			policies, ok := resource.Dynamic[esName]
-			if !ok {
-				continue
-			}
-			mfi, ok := policies[api.MeshFaultInjectionType]
-			if !ok {
-				continue
-			}
-			protocol := util.GetExternalServiceProtocol(es)
-
-			//nolint:staticcheck // SA1019 Zone egress uses old Rules format for external services
-			for _, rule := range mfi.FromRules.Rules {
-				for _, filterChain := range listeners.Egress.FilterChains {
-					if filterChain.Name == names.GetEgressFilterChainName(esName, meshName) {
-						if err := configure(rule, filterChain, protocol); err != nil {
-							return err
-						}
-					}
-				}
-			}
-		}
-	}
-	return nil
-}
-
-func configure(
-	fromRules core_rules.Rules,
-	filterChain *envoy_listener.FilterChain,
-	protocol core_meta.Protocol,
-) error {
-	switch protocol {
-	case core_meta.ProtocolHTTP, core_meta.ProtocolHTTP2, core_meta.ProtocolGRPC:
-		for _, rule := range fromRules {
-			conf := rule.Conf.(api.Conf)
-			from := rule.Subset
-
-			configurer := plugin_xds.LegacyConfigurer{
-				FaultInjections: pointer.Deref(conf.Http),
-				From:            from,
-			}
-
-			if err := configurer.ConfigureHttpListener(filterChain); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
