+++ pkg/plugins/policies/meshfaultinjection/plugin/xds/configurer.go
@@ -2,8 +2,11 @@ package xds
 
 import (
 	envoy_listener "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
+	envoy_filter_fault "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/common/fault/v3"
 	envoy_http_fault "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/fault/v3"
 	envoy_hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
+	envoy_type "github.com/envoyproxy/go-control-plane/envoy/type/v3"
+	"k8s.io/apimachinery/pkg/util/intstr"
 
 	common_api "github.com/kumahq/kuma/v3/api/common/v1alpha1"
 	"github.com/kumahq/kuma/v3/pkg/core/kri"
@@ -105,4 +108,66 @@ func configureFault(fault policies_api.FaultInjectionConf) (*envoy_http_fault.HT
 	faultConfig.ResponseRateLimit = rrl
 
 	return faultConfig, nil
+}
+
+func convertDelay(delay *policies_api.DelayConf) (*envoy_filter_fault.FaultDelay, error) {
+	if delay == nil {
+		return nil, nil
+	}
+	percentage, err := fractionalPercent(delay.Percentage)
+	if err != nil {
+		return nil, err
+	}
+	return &envoy_filter_fault.FaultDelay{
+		FaultDelaySecifier: &envoy_filter_fault.FaultDelay_FixedDelay{FixedDelay: util_proto.Duration(delay.Value.Duration)},
+		Percentage:         percentage,
+	}, nil
+}
+
+func convertAbort(abort *policies_api.AbortConf) (*envoy_http_fault.FaultAbort, error) {
+	if abort == nil {
+		return nil, nil
+	}
+	percentage, err := fractionalPercent(abort.Percentage)
+	if err != nil {
+		return nil, err
+	}
+	return &envoy_http_fault.FaultAbort{
+		ErrorType:  &envoy_http_fault.FaultAbort_HttpStatus{HttpStatus: uint32(abort.HttpStatus)},
+		Percentage: percentage,
+	}, nil
+}
+
+func convertResponseRateLimit(responseBandwidth *policies_api.ResponseBandwidthConf) (*envoy_filter_fault.FaultRateLimit, error) {
+	if responseBandwidth == nil {
+		return nil, nil
+	}
+
+	limitKbps, err := listeners_v3.ConvertBandwidthToKbps(responseBandwidth.Limit)
+	if err != nil {
+		return nil, err
+	}
+
+	percentage, err := fractionalPercent(responseBandwidth.Percentage)
+	if err != nil {
+		return nil, err
+	}
+
+	return &envoy_filter_fault.FaultRateLimit{
+		LimitType: &envoy_filter_fault.FaultRateLimit_FixedLimit_{
+			FixedLimit: &envoy_filter_fault.FaultRateLimit_FixedLimit{
+				LimitKbps: limitKbps,
+			},
+		},
+		Percentage: percentage,
+	}, nil
+}
+
+func fractionalPercent(percentage intstr.IntOrString) (*envoy_type.FractionalPercent, error) {
+	decimal, err := common_api.NewDecimalFromIntOrString(percentage)
+	if err != nil {
+		return nil, err
+	}
+	value, _ := decimal.Float64()
+	return listeners_v3.ConvertPercentage(util_proto.Double(value)), nil
 }
